updated blocking code

+++ Core/Inc/SIMCom.h
@@ -1,5 +1,5 @@
 #include "stm32f1xx_hal.h"
 
 uint8_t SIMCom_handShake(void); //handShake
-uint8_t SIMCom_post(double vol, double bat); //post method
+uint8_t SIMCom_post(float vol, float bat); //post method
 uint8_t SIMCom_Get(void); //get method
